+++ src/test/main/engine/augment/effectDB/StatEffectTest.java
@@ -34,7 +34,7 @@ class StatEffectTest {
 
     @Test
     void moveSpeedIncrease() {
-        int originalSpeed = stats.getMoveSpeed();
+        float originalSpeed = stats.getMoveSpeed();
 
         StatEffect effect = new StatEffect(StatEffect.StatType.SPEED, 1);
         effect.apply(player);
@@ -45,7 +45,7 @@ class StatEffectTest {
 
     @Test
     void bulletSpeedDecrease() {
-        int originalBulletSpeed = stats.getBulletSpeed();
+        float originalBulletSpeed = stats.getBulletSpeed();
 
         StatEffect effect = new StatEffect(StatEffect.StatType.BULLET_SPEED, 3);
         effect.apply(player);
